pyproject.toml
[project]
name = "game_building"
version = "0.1.0"
description = ""
authors = [
    {name = "Abdelrahman Hussein",email = "abdurrahmanhussein500@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "mongoengine (>=0.29.1,<0.30.0)",
    "channels (>=4.2.2,<5.0.0)",
    "channels-redis (>=4.2.0,<5.0.0)",
    "django-mongodb-backend (==5.2.*)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "pyjwt (>=2.10.0,<3.0.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "djangorestframework (>=3.15.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
